+++ packages/orq-rc/RELEASES.md
@@ -7259,3 +7259,13 @@ Based on:
 - [typescript v4.5.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.5 - packages/orq-rc
+
+## 2026-02-25 14:30:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.6 - packages/orq-rc